+++ .idea/.name
@@ -1,1 +1,1 @@
-FirstTry
+CFTS
