+++ app/src/androidTest/java/com/example/cfts/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.firsttry;
+package com.example.cfts;
 
 import android.content.Context;
 
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
     public void useAppContext() {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
-        assertEquals("com.example.firsttry", appContext.getPackageName());
+        assertEquals("com.example.cfts", appContext.getPackageName());
     }
 }
