+++ app/src/main/java/com/example/cfts/Calendar.java
@@ -1,4 +1,4 @@
-package com.example.firsttry;
+package com.example.cfts;
 
 import androidx.appcompat.app.AppCompatActivity;
