+++ app/src/main/java/com/example/cfts/DefaultWorkouts.java
@@ -1,4 +1,4 @@
-package com.example.firsttry;
+package com.example.cfts;
 
 import java.util.ArrayList;
