+++ app/src/main/java/com/example/cfts/History.java
@@ -1,4 +1,4 @@
-package com.example.firsttry;
+package com.example.cfts;
 
 import java.io.Serializable;
