+++ app/src/main/java/com/example/cfts/RowDataDetail.java
@@ -1,4 +1,4 @@
-package com.example.firsttry;
+package com.example.cfts;
 
 public class RowDataDetail {
     private String title;
