+++ app/src/main/java/com/example/cfts/StringFormatter.java
@@ -1,4 +1,4 @@
-package com.example.firsttry;
+package com.example.cfts;
 
 import java.util.ArrayList;
