+++ app/src/main/java/com/example/cfts/TimerActivity.java
@@ -1,4 +1,4 @@
-package com.example.firsttry;
+package com.example.cfts;
 
 import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AppCompatActivity;
