+++ app/src/main/java/com/example/cfts/Workout.java
@@ -1,4 +1,4 @@
-package com.example.firsttry;
+package com.example.cfts;
 
 import java.io.Serializable;
 import java.util.ArrayList;
