+++ app/src/main/java/com/example/cfts/WorkoutActivity.java
@@ -1,4 +1,4 @@
-package com.example.firsttry;
+package com.example.cfts;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
