+++ app/src/main/res/layout/list_exercises.xml
@@ -8,7 +8,6 @@
         android:id="@+id/txttitle"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_toRightOf="@+id/icon"
         android:layout_marginTop="5dp"
         android:gravity="center"
         android:textColor="@color/white"
